fix: namespace refactor & proguard
