+++ features/exploredash/src/main/java/org/dash/wallet/features/exploredash/data/dashspend/piggycards/model/OrderRequest.kt
@@ -14,17 +14,21 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.dash.wallet.features.exploredash.data.piggycards.model
+package org.dash.wallet.features.exploredash.data.dashspend.piggycards.model
 
 import com.google.gson.annotations.SerializedName
 
-data class Location(
-    @SerializedName("id")
-    val id: String,
-    @SerializedName("address")
-    val address: String,
-    @SerializedName("latitude")
-    val latitude: Double,
-    @SerializedName("longitude")
-    val longitude: Double
+data class OrderRequest(
+    @SerializedName("merchantId")
+    val merchantId: String,
+    @SerializedName("amount")
+    val amount: Double,
+    @SerializedName("email")
+    val email: String,
+    @SerializedName("name")
+    val name: String,
+    @SerializedName("ipAddress")
+    val ipAddress: String?,
+    @SerializedName("location")
+    val location: String?
 )
