+++ features/exploredash/src/main/java/org/dash/wallet/features/exploredash/data/dashspend/piggycards/model/OrderResponse.kt
@@ -14,25 +14,15 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.dash.wallet.features.exploredash.data.piggycards.model
+package org.dash.wallet.features.exploredash.data.dashspend.piggycards.model
 
 import com.google.gson.annotations.SerializedName
 
-data class MerchantDetails(
+data class OrderResponse(
     @SerializedName("id")
     val id: String,
-    @SerializedName("name")
-    val name: String,
-    @SerializedName("discount")
-    val discount: Double?,
-    @SerializedName("imageUrl")
-    val imageUrl: String?,
-    @SerializedName("fixedDenominations")
-    val fixedDenominations: List<Double>?,
-    @SerializedName("variableDenomination")
-    val variableDenomination: Boolean,
-    @SerializedName("minAmount")
-    val minAmount: Double?,
-    @SerializedName("maxAmount")
-    val maxAmount: Double?
+    @SerializedName("payTo")
+    val payTo: String,
+    @SerializedName("payMessage")
+    val payMessage: String
 )
