+++ features/exploredash/src/main/java/org/dash/wallet/features/exploredash/data/dashspend/piggycards/model/VerifyOtpResponse.kt
@@ -14,19 +14,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.dash.wallet.features.exploredash.data.piggycards.model
+package org.dash.wallet.features.exploredash.data.dashspend.piggycards.model
 
 import com.google.gson.annotations.SerializedName
 
-data class SignupRequest(
-    @SerializedName("firstName")
-    val firstName: String,
-    @SerializedName("lastName")
-    val lastName: String,
-    @SerializedName("email")
-    val email: String,
-    @SerializedName("country")
-    val country: String,
-    @SerializedName("state")
-    val state: String? = null
+data class VerifyOtpResponse(
+    @SerializedName("generatedPassword")
+    val generatedPassword: String
 )
