+++ wallet/proguard.cfg
@@ -225,11 +225,6 @@
 -keepattributes SourceFile,LineNumberTable
 -keep class kotlin.Metadata { *; }
 
-# CTX
--keep class org.dash.wallet.features.exploredash.data.ctxspend.model.** { *; }
--keep,allowobfuscation interface org.dash.wallet.features.exploredash.network.service.ctxspend.CTXSpendApi
--keep,allowobfuscation interface org.dash.wallet.features.exploredash.network.service.ctxspend.CTXSpendTokenApi
-
 # Google Drive
 -keep class com.google.api.client.** { *; }
 -keep class com.google.api.services.** { *; }
@@ -258,5 +253,11 @@
 -keepnames class com.google.firebase.auth.**
 -keep interface com.google.firebase.auth.** { *; }
 
+# CTX
+-keep class org.dash.wallet.features.exploredash.data.dashspend.ctx.model.** { *; }
+-keep,allowobfuscation interface org.dash.wallet.features.exploredash.network.service.ctxspend.CTXSpendApi
+-keep,allowobfuscation interface org.dash.wallet.features.exploredash.network.service.ctxspend.CTXSpendTokenApi
+
 # PiggyCards
--keep class org.dash.wallet.features.exploredash.data.piggycards.model.** { *; }
+-keep class org.dash.wallet.features.exploredash.data.dashspend.piggycards.model.** { *; }
+-keep,allowobfuscation interface org.dash.wallet.features.exploredash.network.service.piggycards.PiggyCardsApi
